refactor: move section code to new dir
